chore: Rename TableBody2 -> TableBody
